Adds HttpResponse class for handling success and failure responses
This commit adds a new HttpResponse class that provides static methods for handling success and failure responses in the Express framework. The `success` method sets the response status to 200 and sends a JSON object with a `success` property set to true, along with optional `data` and `details` properties. The `failure` method allows setting a custom status code and sends a JSON object with a `success` property set to false, along with a `message` property. This class will help streamline response handling in the codebase.